change GPS time to be correct
The Raspberry pi doesn't have an RTC, so when it's started up after being stopped
for a while, the local time is significantly delayed. The footage should
show the GPS time instead of the local time. If the local time differes
significantly from the GPS time, correct the local time.